Extract wave response contracts
